+++ internal/ghcmd/cmd_test.go
@@ -1,4 +1,4 @@
-package main
+package ghcmd
 
 import (
 	"bytes"
